requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==4.1.13
dnspython==2.6.1
filelock==3.14.0
nulltype==2.3.1
plaid-python==22.0.0
platformdirs==4.2.2
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.26.2
